backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
python-multipart
